+++ lib/models/achievement.freezed.dart
@@ -1,0 +1,371 @@
+// coverage:ignore-file
+// GENERATED CODE - DO NOT MODIFY BY HAND
+// ignore_for_file: type=lint
+// ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides, invalid_annotation_target, unnecessary_question_mark
+
+part of 'achievement.dart';
+
+// **************************************************************************
+// FreezedGenerator
+// **************************************************************************
+
+T _$identity<T>(T value) => value;
+
+final _privateConstructorUsedError = UnsupportedError(
+    'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more information: https://github.com/rrousselGit/freezed#adding-getters-and-methods-to-our-models');
+
+Achievement _$AchievementFromJson(Map<String, dynamic> json) {
+  return _Achievement.fromJson(json);
+}
+
+/// @nodoc
+mixin _$Achievement {
+  String get id => throw _privateConstructorUsedError;
+  String get title => throw _privateConstructorUsedError;
+  String get description => throw _privateConstructorUsedError;
+  String get iconEmoji => throw _privateConstructorUsedError;
+  int get currentValue => throw _privateConstructorUsedError;
+  int get targetValue => throw _privateConstructorUsedError;
+  int get tier => throw _privateConstructorUsedError;
+  int get maxTier => throw _privateConstructorUsedError;
+  int get dropletReward => throw _privateConstructorUsedError;
+  bool get isClaimed => throw _privateConstructorUsedError;
+
+  /// Serializes this Achievement to a JSON map.
+  Map<String, dynamic> toJson() => throw _privateConstructorUsedError;
+
+  /// Create a copy of Achievement
+  /// with the given fields replaced by the non-null parameter values.
+  @JsonKey(includeFromJson: false, includeToJson: false)
+  $AchievementCopyWith<Achievement> get copyWith =>
+      throw _privateConstructorUsedError;
+}
+
+/// @nodoc
+abstract class $AchievementCopyWith<$Res> {
+  factory $AchievementCopyWith(
+          Achievement value, $Res Function(Achievement) then) =
+      _$AchievementCopyWithImpl<$Res, Achievement>;
+  @useResult
+  $Res call(
+      {String id,
+      String title,
+      String description,
+      String iconEmoji,
+      int currentValue,
+      int targetValue,
+      int tier,
+      int maxTier,
+      int dropletReward,
+      bool isClaimed});
+}
+
+/// @nodoc
+class _$AchievementCopyWithImpl<$Res, $Val extends Achievement>
+    implements $AchievementCopyWith<$Res> {
+  _$AchievementCopyWithImpl(this._value, this._then);
+
+  // ignore: unused_field
+  final $Val _value;
+  // ignore: unused_field
+  final $Res Function($Val) _then;
+
+  /// Create a copy of Achievement
+  /// with the given fields replaced by the non-null parameter values.
+  @pragma('vm:prefer-inline')
+  @override
+  $Res call({
+    Object? id = null,
+    Object? title = null,
+    Object? description = null,
+    Object? iconEmoji = null,
+    Object? currentValue = null,
+    Object? targetValue = null,
+    Object? tier = null,
+    Object? maxTier = null,
+    Object? dropletReward = null,
+    Object? isClaimed = null,
+  }) {
+    return _then(_value.copyWith(
+      id: null == id
+          ? _value.id
+          : id // ignore: cast_nullable_to_non_nullable
+              as String,
+      title: null == title
+          ? _value.title
+          : title // ignore: cast_nullable_to_non_nullable
+              as String,
+      description: null == description
+          ? _value.description
+          : description // ignore: cast_nullable_to_non_nullable
+              as String,
+      iconEmoji: null == iconEmoji
+          ? _value.iconEmoji
+          : iconEmoji // ignore: cast_nullable_to_non_nullable
+              as String,
+      currentValue: null == currentValue
+          ? _value.currentValue
+          : currentValue // ignore: cast_nullable_to_non_nullable
+              as int,
+      targetValue: null == targetValue
+          ? _value.targetValue
+          : targetValue // ignore: cast_nullable_to_non_nullable
+              as int,
+      tier: null == tier
+          ? _value.tier
+          : tier // ignore: cast_nullable_to_non_nullable
+              as int,
+      maxTier: null == maxTier
+          ? _value.maxTier
+          : maxTier // ignore: cast_nullable_to_non_nullable
+              as int,
+      dropletReward: null == dropletReward
+          ? _value.dropletReward
+          : dropletReward // ignore: cast_nullable_to_non_nullable
+              as int,
+      isClaimed: null == isClaimed
+          ? _value.isClaimed
+          : isClaimed // ignore: cast_nullable_to_non_nullable
+              as bool,
+    ) as $Val);
+  }
+}
+
+/// @nodoc
+abstract class _$$AchievementImplCopyWith<$Res>
+    implements $AchievementCopyWith<$Res> {
+  factory _$$AchievementImplCopyWith(
+          _$AchievementImpl value, $Res Function(_$AchievementImpl) then) =
+      __$$AchievementImplCopyWithImpl<$Res>;
+  @override
+  @useResult
+  $Res call(
+      {String id,
+      String title,
+      String description,
+      String iconEmoji,
+      int currentValue,
+      int targetValue,
+      int tier,
+      int maxTier,
+      int dropletReward,
+      bool isClaimed});
+}
+
+/// @nodoc
+class __$$AchievementImplCopyWithImpl<$Res>
+    extends _$AchievementCopyWithImpl<$Res, _$AchievementImpl>
+    implements _$$AchievementImplCopyWith<$Res> {
+  __$$AchievementImplCopyWithImpl(
+      _$AchievementImpl _value, $Res Function(_$AchievementImpl) _then)
+      : super(_value, _then);
+
+  /// Create a copy of Achievement
+  /// with the given fields replaced by the non-null parameter values.
+  @pragma('vm:prefer-inline')
+  @override
+  $Res call({
+    Object? id = null,
+    Object? title = null,
+    Object? description = null,
+    Object? iconEmoji = null,
+    Object? currentValue = null,
+    Object? targetValue = null,
+    Object? tier = null,
+    Object? maxTier = null,
+    Object? dropletReward = null,
+    Object? isClaimed = null,
+  }) {
+    return _then(_$AchievementImpl(
+      id: null == id
+          ? _value.id
+          : id // ignore: cast_nullable_to_non_nullable
+              as String,
+      title: null == title
+          ? _value.title
+          : title // ignore: cast_nullable_to_non_nullable
+              as String,
+      description: null == description
+          ? _value.description
+          : description // ignore: cast_nullable_to_non_nullable
+              as String,
+      iconEmoji: null == iconEmoji
+          ? _value.iconEmoji
+          : iconEmoji // ignore: cast_nullable_to_non_nullable
+              as String,
+      currentValue: null == currentValue
+          ? _value.currentValue
+          : currentValue // ignore: cast_nullable_to_non_nullable
+              as int,
+      targetValue: null == targetValue
+          ? _value.targetValue
+          : targetValue // ignore: cast_nullable_to_non_nullable
+              as int,
+      tier: null == tier
+          ? _value.tier
+          : tier // ignore: cast_nullable_to_non_nullable
+              as int,
+      maxTier: null == maxTier
+          ? _value.maxTier
+          : maxTier // ignore: cast_nullable_to_non_nullable
+              as int,
+      dropletReward: null == dropletReward
+          ? _value.dropletReward
+          : dropletReward // ignore: cast_nullable_to_non_nullable
+              as int,
+      isClaimed: null == isClaimed
+          ? _value.isClaimed
+          : isClaimed // ignore: cast_nullable_to_non_nullable
+              as bool,
+    ));
+  }
+}
+
+/// @nodoc
+@JsonSerializable()
+class _$AchievementImpl extends _Achievement {
+  const _$AchievementImpl(
+      {required this.id,
+      required this.title,
+      required this.description,
+      required this.iconEmoji,
+      required this.currentValue,
+      required this.targetValue,
+      this.tier = 1,
+      this.maxTier = 3,
+      this.dropletReward = 15,
+      this.isClaimed = false})
+      : super._();
+
+  factory _$AchievementImpl.fromJson(Map<String, dynamic> json) =>
+      _$$AchievementImplFromJson(json);
+
+  @override
+  final String id;
+  @override
+  final String title;
+  @override
+  final String description;
+  @override
+  final String iconEmoji;
+  @override
+  final int currentValue;
+  @override
+  final int targetValue;
+  @override
+  @JsonKey()
+  final int tier;
+  @override
+  @JsonKey()
+  final int maxTier;
+  @override
+  @JsonKey()
+  final int dropletReward;
+  @override
+  @JsonKey()
+  final bool isClaimed;
+
+  @override
+  String toString() {
+    return 'Achievement(id: $id, title: $title, description: $description, iconEmoji: $iconEmoji, currentValue: $currentValue, targetValue: $targetValue, tier: $tier, maxTier: $maxTier, dropletReward: $dropletReward, isClaimed: $isClaimed)';
+  }
+
+  @override
+  bool operator ==(Object other) {
+    return identical(this, other) ||
+        (other.runtimeType == runtimeType &&
+            other is _$AchievementImpl &&
+            (identical(other.id, id) || other.id == id) &&
+            (identical(other.title, title) || other.title == title) &&
+            (identical(other.description, description) ||
+                other.description == description) &&
+            (identical(other.iconEmoji, iconEmoji) ||
+                other.iconEmoji == iconEmoji) &&
+            (identical(other.currentValue, currentValue) ||
+                other.currentValue == currentValue) &&
+            (identical(other.targetValue, targetValue) ||
+                other.targetValue == targetValue) &&
+            (identical(other.tier, tier) || other.tier == tier) &&
+            (identical(other.maxTier, maxTier) || other.maxTier == maxTier) &&
+            (identical(other.dropletReward, dropletReward) ||
+                other.dropletReward == dropletReward) &&
+            (identical(other.isClaimed, isClaimed) ||
+                other.isClaimed == isClaimed));
+  }
+
+  @JsonKey(includeFromJson: false, includeToJson: false)
+  @override
+  int get hashCode => Object.hash(
+      runtimeType,
+      id,
+      title,
+      description,
+      iconEmoji,
+      currentValue,
+      targetValue,
+      tier,
+      maxTier,
+      dropletReward,
+      isClaimed);
+
+  /// Create a copy of Achievement
+  /// with the given fields replaced by the non-null parameter values.
+  @JsonKey(includeFromJson: false, includeToJson: false)
+  @override
+  @pragma('vm:prefer-inline')
+  _$$AchievementImplCopyWith<_$AchievementImpl> get copyWith =>
+      __$$AchievementImplCopyWithImpl<_$AchievementImpl>(this, _$identity);
+
+  @override
+  Map<String, dynamic> toJson() {
+    return _$$AchievementImplToJson(
+      this,
+    );
+  }
+}
+
+abstract class _Achievement extends Achievement {
+  const factory _Achievement(
+      {required final String id,
+      required final String title,
+      required final String description,
+      required final String iconEmoji,
+      required final int currentValue,
+      required final int targetValue,
+      final int tier,
+      final int maxTier,
+      final int dropletReward,
+      final bool isClaimed}) = _$AchievementImpl;
+  const _Achievement._() : super._();
+
+  factory _Achievement.fromJson(Map<String, dynamic> json) =
+      _$AchievementImpl.fromJson;
+
+  @override
+  String get id;
+  @override
+  String get title;
+  @override
+  String get description;
+  @override
+  String get iconEmoji;
+  @override
+  int get currentValue;
+  @override
+  int get targetValue;
+  @override
+  int get tier;
+  @override
+  int get maxTier;
+  @override
+  int get dropletReward;
+  @override
+  bool get isClaimed;
+
+  /// Create a copy of Achievement
+  /// with the given fields replaced by the non-null parameter values.
+  @override
+  @JsonKey(includeFromJson: false, includeToJson: false)
+  _$$AchievementImplCopyWith<_$AchievementImpl> get copyWith =>
+      throw _privateConstructorUsedError;
+}