+++ lib/providers/achievement_provider.g.dart
@@ -1,0 +1,27 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+part of 'achievement_provider.dart';
+
+// **************************************************************************
+// RiverpodGenerator
+// **************************************************************************
+
+String _$achievementNotifierHash() =>
+    r'1ea49cc89898429dcae955236d7c6ca63264af94';
+
+/// See also [AchievementNotifier].
+@ProviderFor(AchievementNotifier)
+final achievementNotifierProvider = AutoDisposeAsyncNotifierProvider<
+    AchievementNotifier, List<Achievement>>.internal(
+  AchievementNotifier.new,
+  name: r'achievementNotifierProvider',
+  debugGetCreateSourceHash: const bool.fromEnvironment('dart.vm.product')
+      ? null
+      : _$achievementNotifierHash,
+  dependencies: null,
+  allTransitiveDependencies: null,
+);
+
+typedef _$AchievementNotifier = AutoDisposeAsyncNotifier<List<Achievement>>;
+// ignore_for_file: type=lint
+// ignore_for_file: subtype_of_sealed_class, invalid_use_of_internal_member, invalid_use_of_visible_for_testing_member, deprecated_member_use_from_same_package